requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.1
certifi==2019.9.11
dj-database-url==0.5.0
Django==3.1
django-appconf==1.0.4
django-bootstrap4==2.2.0
django-bootstrap4-datetimepicker==4.2
django-bootstrap4-form==4.0.1
django-bootstrap4-nav==0.0.1
django-ckeditor==5.9.0
django-ckeditor-filebrowser-filer==0.3.0
django-ckeditor-filer==1.2.0
django-compressor==2.4
django-filer==1.7.1
django-js-asset==1.2.2
django-libsass==0.8
django-mptt==0.11.0
django-mysql-pymysql==0.1
django-polymorphic==2.1.2
easy-thumbnails==2.7
importlib-metadata==1.7.0
libsass==0.20.0
Pillow==7.2.0
pipenv==2018.11.26
protobuf==3.12.4
psycopg2-binary==2.8.5
pytz==2020.1
rcssmin==1.0.6
rjsmin==1.1.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
Unidecode==1.1.1
virtualenv==16.7.7
virtualenv-clone==0.5.3
whitenoise==5.2.0
zipp==3.1.0
